requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
anthropic==0.25.0
supabase==2.4.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.25.2
